Updated file headers.

+++ Noise.h
@@ -1,3 +1,21 @@
+/*
+ * Torch
+ * Copyright (C) 2015 Jason Coon
+ * 
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 #define MAX_DIMENSION ((MATRIX_WIDTH > MATRIX_HEIGHT) ? MATRIX_WIDTH : MATRIX_HEIGHT)
 
 // The 16 bit version of our coordinates
@@ -9,9 +27,9 @@ static uint16_t noisez;
 // use the z-axis for "time".  speed determines how fast time moves forward.  Try
 // 1 for a very slow moving effect, or 60 for something that ends up looking like
 // water.
-uint16_t noisespeedx = 1;
-uint16_t noisespeedy = 1;
-uint16_t noisespeedz = 1;
+uint32_t noisespeedx = 1;
+uint32_t noisespeedy = 1;
+uint32_t noisespeedz = 1;
 
 // Scale determines how far apart the pixels in our noise matrix are.  Try
 // changing these values around to see how it affects the motion of the display.  The
